Refactor headless module, introduce world
