+++ maplibre/src/headless/window.rs
@@ -1,0 +1,23 @@
+use crate::window::{MapWindow, MapWindowConfig, WindowSize};
+
+pub struct HeadlessMapWindowConfig {
+    pub size: WindowSize,
+}
+
+impl MapWindowConfig for HeadlessMapWindowConfig {
+    type MapWindow = HeadlessMapWindow;
+
+    fn create(&self) -> Self::MapWindow {
+        Self::MapWindow { size: self.size }
+    }
+}
+
+pub struct HeadlessMapWindow {
+    size: WindowSize,
+}
+
+impl MapWindow for HeadlessMapWindow {
+    fn size(&self) -> WindowSize {
+        self.size
+    }
+}
